# docs/swagger.json
{"openapi":"3.0.0","paths":{"/":{"get":{"operationId":"AppController_ping","parameters":[],"responses":{"200":{"description":"pong"}}}},"/api/crm/v1/customers":{"get":{"operationId":"CustomersController_findById","parameters":[{"name":"id","required":true,"in":"query","schema":{"type":"number"}}],"responses":{"200":{"description":"Get all customers.","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Customer"}}}}}},"tags":["Customers"]}},"/api/crm/v1/activities":{"get":{"operationId":"ActivitiesController_findAll","parameters":[],"responses":{"200":{"description":"Get all activities.","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Activity"}}}}}},"tags":["Activities"]},"post":{"operationId":"ActivitiesController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateActivityDto"}}}},"responses":{"201":{"description":"Create a new activity.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Activity"}}}}},"tags":["Activities"]}},"/api/crm/v1/subscriptions":{"post":{"operationId":"SubscriptionsController_createSubscription","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateSubscriptionDto"}}}},"responses":{"201":{"description":""}},"tags":["Subscriptions"]}}},"info":{"title":"CRM API","description":"CRM for WLC","version":"1.0.0","contact":{"name":"Ruben L","url":null,"email":"rubenlellouche26@gmail.com"}},"tags":[{"name":"CRM","description":""}],"servers":[{"url":"https://webleadercomprd.herokuapp.com","description":"Production Server"}],"components":{"schemas":{"Customer":{"type":"object","properties":{"id":{"type":"number","description":"Customer ID"},"nom":{"type":"string","description":"Customer name"},"mail":{"type":"string","description":"Customer email"}},"required":["id","nom","mail"]},"Activity":{"type":"object","properties":{"id":{"type":"number","example":1,"description":"Customer ID"},"title":{"type":"string","example":"John Doe","description":"Customer name"},"description":{"type":"string","example":"John Doe","description":"Customer name"}}},"CreateActivityDto":{"type":"object","properties":{"title":{"type":"string"},"description":{"type":"string"},"customers":{"type":"array","items":{"type":"string"}}},"required":["title","description"]},"CreateSubscriptionDto":{"type":"object","properties":{}}}}}